Add files via upload
This soccer database comes from Kaggle and is well suited for data analysis of the european football league. It contains data for soccer matches, players, and teams from several European countries from 2008 to 2016. 
In this project, we will explore and analyze the dataset to gain insights and interesting stats about players, teams, and leagues.